% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dose_Combination_Code.R
\name{Dist_restnull2}
\alias{Dist_restnull2}
\title{\code{Dist_restnull2} This function evaluates the p value under the least favourable null approach}
\usage{
Dist_restnull2(levelsDrugA, levelsDrugB, len, ind, indcomb, Crit)
}
\arguments{
\item{levelsDrugA}{no of dose groups for Drug A (r)}

\item{levelsDrugB}{no of dose groups for Drug B (s)}

\item{len}{Variable contain the number of patients in each dose combination group}

\item{ind}{all the drug dose combination groups (index including monotherapies and placebo control) in our study}

\item{indcomb}{all the drug dose combination groups (index excluding monotherapies and placebo control) in our study}

\item{Crit}{value of obs test statistics where the p value should be computed}
}
\value{
returns the p-value vector
}
\description{
\code{Dist_restnull2} This function evaluates the p value under the least favourable null approach
}
\references{
\url{http://stackoverflow.com/a/5577647}
}

