% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dose_Combination_Code.R
\name{eval_f}
\alias{eval_f}
\title{\code{eval_f} This function is for computing the SSE while fitting the model}
\usage{
eval_f(Y, len, x)
}
\arguments{
\item{Y}{obsservation for response}

\item{len}{Variable contain the number of patients in each dose combination group}

\item{x}{mean estimates under each dose group}
}
\value{
the sum
}
\description{
\code{eval_f} This function is for computing the SSE while fitting the model
}

